feat: adds fileReader module for reading wallets file
This commit adds a new module called fileReader.ts that contains a function
readWalletsFile(). This function reads a file named "wallets.txt" and returns
an array of wallet addresses. The file is read line by line using the
readline module and the fs module is used to create a readable stream. The
function resolves with the array of wallet addresses when the file reading is
finished. If there is an error during the file reading, it rejects with false.

The index.ts file is also updated to use the readWalletsFile() function from
the fileReader module. The previous code that read the file using readline and
fs modules is replaced with a single line of code that calls the
readWalletsFile() function. The returned array of wallet addresses is assigned
to the wallets variable.